refactor: improved x5c-package structure
