+++ docker-compose/wallet-api/config/_features.conf
@@ -1,3 +1,5 @@
 enabledFeatures = [
-
+    # trusted-ca
+    # entra,
+    # ...
 ]
