+++ waltid-services/waltid-wallet-api/config/_features.conf
@@ -1,4 +1,5 @@
 enabledFeatures = [
+    # trusted-ca
     # entra,
     # ...
 ]
